move playbook and config parser to separate file
